Move TrueHashPlugin into the webpack5-localization-plugin package.
