+++ common/config/rush/repo-state.json
@@ -1,5 +1,5 @@
 // DO NOT MODIFY THIS FILE MANUALLY BUT DO COMMIT IT. It is generated and used by Rush.
 {
-  "pnpmShrinkwrapHash": "4a56584d543c9c5a5d512d77ac32785fc025f9ae",
+  "pnpmShrinkwrapHash": "4c17c55360a1f1c707f1aa08e0089c7563d471c0",
   "preferredVersionsHash": "40d4640a94cff77f7808a2f1960cc76231eb6f86"
 }
