+++ common/reviews/api/webpack5-localization-plugin.api.md
@@ -4,17 +4,31 @@
 
 ```ts
 
+/// <reference types="node" />
+
 import type { Chunk } from 'webpack';
 import type { Compiler } from 'webpack';
+import type { default as default_2 } from 'webpack';
 import { ILocalizationFile } from '@rushstack/localization-utilities';
 import type { IPseudolocaleOptions } from '@rushstack/localization-utilities';
 import type { LoaderContext } from 'webpack';
 import type { WebpackPluginInstance } from 'webpack';
 
 // @public (undocumented)
+export interface ICustomHashFunctionOptions extends ITrueHashPluginOptionsBase {
+    hashFunction: (contents: string | Buffer) => string;
+}
+
+// @public (undocumented)
 export interface IDefaultLocaleOptions {
     fillMissingTranslationStrings?: boolean;
     localeName: string;
+}
+
+// @public (undocumented)
+export interface IHashAlgorithmOptions extends ITrueHashPluginOptionsBase {
+    // Warning: (ae-forgotten-export) The symbol "WebpackHash" needs to be exported by the entry point index.d.ts
+    hash?: WebpackHash;
 }
 
 // @public (undocumented)
@@ -125,6 +139,15 @@ export interface _IStringPlaceholder {
     valuesByLocale: Map<string, string>;
 }
 
+// @public (undocumented)
+export type ITrueHashPluginOptions = IHashAlgorithmOptions | ICustomHashFunctionOptions;
+
+// @public (undocumented)
+export interface ITrueHashPluginOptionsBase {
+    // (undocumented)
+    stageOverride?: number;
+}
+
 // @public
 export class LocalizationPlugin implements WebpackPluginInstance {
     constructor(options: ILocalizationPluginOptions);
@@ -137,6 +160,12 @@ export class LocalizationPlugin implements WebpackPluginInstance {
     getPlaceholder(localizedFileKey: string, stringName: string): _IStringPlaceholder | undefined;
     // (undocumented)
     readonly stringKeys: Map<string, _IStringPlaceholder>;
+}
+
+// @public (undocumented)
+export class TrueHashPlugin implements WebpackPluginInstance {
+    constructor(options: ITrueHashPluginOptions);
+    apply(compiler: Compiler): void;
 }
 
 // (No @packageDocumentation comment for this package)
