--- common/reviews/api/webpack5-true-hash-plugin.api.md
@@ -1,41 +1,0 @@
-## API Report File for "@rushstack/webpack5-true-hash-plugin"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-/// <reference types="node" />
-
-import type { Compiler } from 'webpack';
-import type { default as default_2 } from 'webpack';
-import type { WebpackPluginInstance } from 'webpack';
-
-// @public (undocumented)
-export interface ICustomHashFunctionOptions extends ITrueHashPluginOptionsBase {
-    hashFunction: (contents: string | Buffer) => string;
-}
-
-// @public (undocumented)
-export interface IHashAlgorithmOptions extends ITrueHashPluginOptionsBase {
-    // Warning: (ae-forgotten-export) The symbol "WebpackHash" needs to be exported by the entry point index.d.ts
-    hash?: WebpackHash;
-}
-
-// @public (undocumented)
-export type ITrueHashPluginOptions = IHashAlgorithmOptions | ICustomHashFunctionOptions;
-
-// @public (undocumented)
-export interface ITrueHashPluginOptionsBase {
-    // (undocumented)
-    stageOverride?: number;
-}
-
-// @public (undocumented)
-export class TrueHashPlugin implements WebpackPluginInstance {
-    constructor(options: ITrueHashPluginOptions);
-    apply(compiler: Compiler): void;
-}
-
-// (No @packageDocumentation comment for this package)
-
-```
