+++ rush.json
@@ -1300,12 +1300,6 @@
       "projectFolder": "webpack/webpack5-localization-plugin",
       "reviewCategory": "libraries",
       "shouldPublish": true
-    },
-    {
-      "packageName": "@rushstack/webpack5-true-hash-plugin",
-      "projectFolder": "webpack/webpack5-true-hash-plugin",
-      "reviewCategory": "libraries",
-      "shouldPublish": true
     }
   ]
 }
