+++ webpack/webpack5-localization-plugin/src/index.ts
@@ -1,7 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-export { LocalizationPlugin, IStringPlaceholder as _IStringPlaceholder } from './LocalizationPlugin';
+export { LocalizationPlugin, type IStringPlaceholder as _IStringPlaceholder } from './LocalizationPlugin';
+export { TrueHashPlugin } from './TrueHashPlugin';
 
 export {
   IDefaultLocaleOptions,
@@ -18,7 +19,11 @@ export {
   ILocalizedStrings,
   IPassthroughLocaleOptions,
   IPseudolocalesOptions,
-  IResolvedMissingTranslations
+  IResolvedMissingTranslations,
+  ITrueHashPluginOptions,
+  ICustomHashFunctionOptions,
+  IHashAlgorithmOptions,
+  ITrueHashPluginOptionsBase
 } from './interfaces';
 
 export { ILocalizedWebpackChunk } from './webpackInterfaces';
