+++ webpack/webpack5-localization-plugin/src/test/LocalizedNoAsync.test.ts
@@ -10,6 +10,7 @@ import { Volume } from 'memfs/lib/volume';
 
 import { LocalizationPlugin } from '../LocalizationPlugin';
 import type { ILocalizationPluginOptions } from '../interfaces';
+import { TrueHashPlugin } from '../TrueHashPlugin';
 import { MemFSPlugin } from './MemFSPlugin';
 
 async function testLocalizedNoAsyncInner(minimize: boolean): Promise<void> {
@@ -26,6 +27,8 @@ async function testLocalizedNoAsyncInner(minimize: boolean): Promise<void> {
     },
     '/'
   );
+
+  const trueHashPlugin: TrueHashPlugin = new TrueHashPlugin({});
 
   const resJsonLoader: string = resolve(__dirname, '../loaders/resjson-loader.js');
   const options: ILocalizationPluginOptions = {
@@ -65,7 +68,7 @@ async function testLocalizedNoAsyncInner(minimize: boolean): Promise<void> {
     },
     output: {
       path: '/release',
-      filename: '[name]-[locale].js'
+      filename: '[name]-[locale]-[contenthash].js'
     },
     module: {
       rules: [
@@ -85,7 +88,7 @@ async function testLocalizedNoAsyncInner(minimize: boolean): Promise<void> {
     },
     context: '/',
     mode: 'production',
-    plugins: [localizationPlugin, new MemFSPlugin(memoryFileSystem)]
+    plugins: [localizationPlugin, trueHashPlugin, new MemFSPlugin(memoryFileSystem)]
   });
 
   const stats: Stats | undefined = await promisify(compiler.run.bind(compiler))();
