+++ webpack/webpack5-localization-plugin/src/test/MixedAsync.test.ts
@@ -10,7 +10,7 @@ import { Volume } from 'memfs/lib/volume';
 
 import { LocalizationPlugin } from '../LocalizationPlugin';
 import type { ILocalizationPluginOptions, ILocalizationStats } from '../interfaces';
-// import { TrueHashPlugin } from '../../../webpack5-true-hash-plugin/src/TrueHashPlugin';
+import { TrueHashPlugin } from '../TrueHashPlugin';
 import { MemFSPlugin } from './MemFSPlugin';
 
 async function testMixedAsyncInner(minimize: boolean): Promise<void> {
@@ -29,7 +29,7 @@ async function testMixedAsyncInner(minimize: boolean): Promise<void> {
     '/'
   );
 
-  // const trueHashPlugin: TrueHashPlugin = new TrueHashPlugin({});
+  const trueHashPlugin: TrueHashPlugin = new TrueHashPlugin({});
 
   let localizationStats: ILocalizationStats | undefined;
   function statsCallback(stats: ILocalizationStats): void {
@@ -87,7 +87,7 @@ async function testMixedAsyncInner(minimize: boolean): Promise<void> {
     },
     context: '/',
     mode: 'production',
-    plugins: [localizationPlugin, new MemFSPlugin(memoryFileSystem)]
+    plugins: [localizationPlugin, trueHashPlugin, new MemFSPlugin(memoryFileSystem)]
   });
 
   const stats: Stats | undefined = await promisify(compiler.run.bind(compiler))();
