--- webpack/webpack5-true-hash-plugin/README.md
@@ -1,9 +1,0 @@
-# @rushstack/webpack5-true-hash-plugin
-
-## Installation
-
-`npm install @rushstack/webpack5-true-hash-plugin --save-dev`
-
-## Overview
-
-This plugin updates the hashes of the output files to be the true content hash of the file. It also supports chunks produced by the `@rushstack/webpack5-localization-plugin`.
