--- webpack/webpack5-true-hash-plugin/config/api-extractor.json
@@ -1,19 +1,0 @@
-{
-  "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
-
-  "mainEntryPointFilePath": "<projectFolder>/lib/index.d.ts",
-
-  "apiReport": {
-    "enabled": true,
-    "reportFolder": "../../../common/reviews/api"
-  },
-
-  "docModel": {
-    "enabled": false,
-    "apiJsonFilePath": "../../../common/temp/api/<unscopedPackageName>.api.json"
-  },
-
-  "dtsRollup": {
-    "enabled": true
-  }
-}
