--- webpack/webpack5-true-hash-plugin/config/jest.config.json
@@ -1,5 +1,0 @@
-{
-  "extends": "local-node-rig/profiles/default/config/jest.config.json",
-
-  "testTimeout": 1e5
-}
