--- webpack/webpack5-true-hash-plugin/config/rig.json
@@ -1,7 +1,0 @@
-{
-  // The "rig.json" file directs tools to look for their config files in an external package.
-  // Documentation for this system: https://www.npmjs.com/package/@rushstack/rig-package
-  "$schema": "https://developer.microsoft.com/json-schemas/rig-package/rig.schema.json",
-
-  "rigPackageName": "local-node-rig"
-}
