--- webpack/webpack5-true-hash-plugin/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@rushstack/webpack5-true-hash-plugin",
-  "version": "0.0.0",
-  "description": "This plugin facilitates true-content hashes of webpack chunks, including support for the @rushstack localization plugins.",
-  "main": "lib/index.js",
-  "typings": "dist/webpack5-true-hash-plugin.d.ts",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/microsoft/rushstack.git",
-    "directory": "webpack/webpack5-true-hash-plugin"
-  },
-  "scripts": {
-    "build": "heft build --clean",
-    "_phase:build": "heft run --only build -- --clean",
-    "_phase:test": "heft run --only test -- --clean"
-  },
-  "peerDependencies": {
-    "webpack": "^5.68.0",
-    "@types/node": "*"
-  },
-  "dependencies": {
-    "@rushstack/node-core-library": "workspace:*"
-  },
-  "devDependencies": {
-    "@rushstack/heft": "workspace:*",
-    "local-node-rig": "workspace:*",
-    "memfs": "3.4.3",
-    "webpack": "~5.82.1",
-    "@rushstack/webpack5-localization-plugin": "workspace:*"
-  },
-  "peerDependenciesMeta": {
-    "@types/node": {
-      "optional": true
-    }
-  }
-}
