--- webpack/webpack5-true-hash-plugin/src/index.ts
@@ -1,10 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
-// See LICENSE in the project root for license information.
-
-export {
-  TrueHashPlugin,
-  type ITrueHashPluginOptions,
-  type ICustomHashFunctionOptions,
-  type IHashAlgorithmOptions,
-  type ITrueHashPluginOptionsBase
-} from './TrueHashPlugin';
