--- webpack/webpack5-true-hash-plugin/src/test/LocalizedRuntime.test.ts
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
-// See LICENSE in the project root for license information.
-
-jest.disableAutomock();
-
-import { runTests } from './LocalizedRuntimeTestBase';
-
-runTests();
