--- webpack/webpack5-true-hash-plugin/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "./node_modules/local-node-rig/profiles/default/tsconfig-base.json",
-  "compilerOptions": {
-    "target": "ES2019"
-  }
-}
